+++ src/common/compandor.h
@@ -15,7 +15,7 @@ typedef struct compandor {
 	} e;
 } compandor_t;
 
-void init_compandor(compandor_t *state, int samplerate, double attack_ms, double recovery_ms, int unaffected_level);
+void init_compandor(compandor_t *state, int samplerate, double attack_ms, double recovery_ms, double unaffected_level);
 void compress_audio(compandor_t *state, sample_t *samples, int num);
 void expand_audio(compandor_t *state, sample_t *samples, int num);
